refactor: add validation for deleting non-pending orders and completed payments

+++ app/Actions/Admin/Order/DeleteOrderAction.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Actions\Admin\Order;
 
+use App\Enums\Order\OrderStatusEnum;
 use App\Models\Order;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Validation\ValidationException;
@@ -16,12 +17,18 @@ final readonly class DeleteOrderAction
     public function handle(Order $order): void
     {
         DB::transaction(function () use ($order): void {
+            if ($order->status !== OrderStatusEnum::PENDING){
+                throw ValidationException::withMessages([
+                    'order' => __('messages.order.cannot_delete_non_pending_order')
+                ]);
+            }
             $order->load('payments');
             if ($order->payments->isNotEmpty() && $order->payments->sum('amount') > 0) {
                 throw ValidationException::withMessages([
                    'order' => __('messages.order.cannot_delete_order_with_payments', ['order_id' => $order->increment_id])
                 ]);
             }
+
 
             $order->delete();
         });
